BT: Fix publication script.

+++ scripts/setup-aapanel/auto/zip.sh
@@ -28,10 +28,12 @@ if [[ -z $GITHUB_ACTIONS ]]; then
   if [[ $? -ne 0 ]]; then echo "Setup source remote failed"; exit 1; fi
 fi
 
-cd $TMP_DIR/source/srs-cloud && git reset --hard HEAD~10 >/dev/null && git pull | grep files &&
-git branch -vv |grep '*' &&
-echo "Cache at $TMP_DIR/source/srs-cloud"
-if [[ $? -ne 0 ]]; then echo "Cache source failed"; exit 1; fi
+if [[ -z $GITHUB_ACTIONS ]]; then
+  cd $TMP_DIR/source/srs-cloud && git reset --hard HEAD~10 >/dev/null && git pull | grep files &&
+  git branch -vv |grep '*' &&
+  echo "Cache at $TMP_DIR/source/srs-cloud"
+  if [[ $? -ne 0 ]]; then echo "Cache source failed"; exit 1; fi
+fi
 
 if [[ -z $GITHUB_ACTIONS ]]; then
   cd $TMP_DIR/source/srs-cloud && git remote set-url origin https://gthub.com/ossrs/srs-cloud.git
